requirements.txt
distro==1.7.0
numpy==1.23.3
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
plotly==5.11.0
pyparsing==3.0.9
PyPDF2==2.11.0
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
tabula-py==2.5.1
tenacity==8.1.0
typing-extensions==4.4.0
